Kraken: move visitors to a separate file
So it can be inlined
